+++ changelog.md
@@ -5,6 +5,7 @@
 ### 1.6.0, soon...
 
 * changed: only record the From name as part of the From field when it isn't empty
+* fixed: purge of old logs will run with default 30 days limit if setting has not been saved
 * added: Sender, CC, BCC columns in list of email logs
 * added: record Reply-To addresses and show on detailed log